add new feature + update some code to make it more robust and OOP + add new repo
# new features:
- #1 WIP moved some code around to make it more OOP of sorts
- #2  WIP maybe should create a maven plugin to manage my app versions
and will make this avilable.

>   *idea:* a maven plugin that runs at compile time traverse the
> project directory and finds the pom.xml and parses using some XML
> parseer and finds the version tag and then puts the version from this
> plguin.

- add a new repository pattern in the pattern-component that will serve
as a basic repository that all **table** must implement and we are still
working on getting a SQL builder that is generic enough to work on
everything thinking about using a [visitor
pattern](https://github.com/mkcoder/visitor-design-pattern-in-java)
along side some dependency injection and builder to create an SQL
Statement the [Spring JDBC
(link)](http://docs.spring.io/spring/docs/current/spring-framework-reference/html/jdbc.html)
we will see

----------
# update
- update the table iterator to have a abstract iterator that every
database table should implement and this will iterate through the model
and retrieve the fields, **meaning these should match the name of the
database column, a work around for this is create various annotations
and check if the annotation is available.  **